requirements.txt
azure-functions==1.7.0
bson==0.5.10
pymongo==3.11.3
python-dateutil==2.8.1
six==1.15.0
